requirements.txt
streamlit==1.45.1
paddleocr==2.8.1
paddlepaddle==2.6.2
opencv-python-headless==4.10.0.84
pillow==11.1.0
numpy==1.26.4
